+++ SharePointFramework/shared-library/src/components/FieldContainer/FieldContainer.module.scss
@@ -1,0 +1,3 @@
+.fieldContainer {
+    margin: 0;
+}
