+++ SharePointFramework/shared-library/src/components/FieldContainer/FieldContainer.tsx
@@ -1,0 +1,21 @@
+import React, { FC } from 'react'
+import styles from './FieldContainer.module.scss'
+import { IFieldContainerProps } from './types'
+import { Field } from '@fluentui/react-components'
+import { IconLabel } from './IconLabel'
+
+export const FieldContainer: FC<IFieldContainerProps> = (props) => {
+  let label = props.label
+  if (props.iconName) {
+    label = {
+      children: () => <IconLabel {...props} />
+    }
+  }
+  return (
+    <div className={styles.fieldContainer}>
+      <Field label={label} required={props.required} hint={props.description ?? props.hint}>
+        {props.children}
+      </Field>
+    </div>
+  )
+}
