+++ SharePointFramework/shared-library/src/components/FieldContainer/IconLabel/IconLabel.module.scss
@@ -1,0 +1,7 @@
+.iconLabel {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 8px;
+    padding-bottom: 5px;
+}
