+++ SharePointFramework/shared-library/src/components/FieldContainer/IconLabel/IconLabel.tsx
@@ -1,0 +1,16 @@
+import { Label } from '@fluentui/react-components'
+import React, { FC } from 'react'
+import { getFluentIcon } from '../../../icons'
+import { IFieldContainerProps } from '../types'
+import styles from './IconLabel.module.scss'
+
+export const IconLabel: FC<IFieldContainerProps> = (props) => {
+  return (
+    <div className={styles.iconLabel}>
+      {getFluentIcon(props.iconName)}
+      <Label weight='semibold' required={props.required}>
+        {props.label}
+      </Label>
+    </div>
+  )
+}
