+++ SharePointFramework/shared-library/src/components/FieldContainer/IconLabel/index.ts
@@ -1,0 +1,1 @@
+export * from './IconLabel'
