+++ SharePointFramework/shared-library/src/components/FieldContainer/index.ts
@@ -1,0 +1,2 @@
+export * from './FieldContainer'
+export * from './types'
