+++ SharePointFramework/shared-library/src/components/FieldContainer/types.ts
@@ -1,0 +1,14 @@
+import { FieldProps } from '@fluentui/react-components'
+import { FluentIconName } from '../../icons/types'
+
+export interface IFieldContainerProps extends FieldProps {
+  /**
+   * The name of the Fluent icon to display.
+   */
+  iconName?: FluentIconName
+
+  /**
+   * Description of the field.
+   */
+  description?: string
+}
