+++ SharePointFramework/shared-library/src/components/index.ts
@@ -10,3 +10,5 @@ export * from './ProjectTimeline'
 export * from './UserMessage'
 export * from './Toolbar'
 export * from './WebPartTitle'
+export * from './FieldContainer'
+export * from './ThemedComponent'
