+++ SharePointFramework/shared-library/src/icons/iconCatalog.ts
@@ -1,0 +1,140 @@
+import {
+  AppsListFilled,
+  AppsListRegular,
+  GridFilled,
+  GridRegular,
+  TextBulletListLtrFilled,
+  TextBulletListLtrRegular,
+  TextSortAscendingFilled,
+  TextSortAscendingRegular,
+  TextSortDescendingFilled,
+  TextSortDescendingRegular,
+  TableSettingsRegular,
+  TableSettingsFilled,
+  GroupListRegular,
+  GroupListFilled,
+  TableCellEditRegular,
+  TableCellEditFilled,
+  ArrowLeftRegular,
+  ArrowLeftFilled,
+  ArrowRightRegular,
+  ArrowRightFilled,
+  EyeRegular,
+  EyeFilled,
+  AddRegular,
+  AddFilled,
+  ArrowSort24Regular,
+  ArrowSort24Filled,
+  ArrowSyncRegular,
+  ArrowSyncFilled,
+  ConvertRangeRegular,
+  ConvertRangeFilled,
+  CheckmarkRegular,
+  CheckmarkFilled,
+  CheckmarkCircleRegular,
+  CheckmarkCircleFilled,
+  PersonRegular,
+  PersonFilled,
+  TextNumberFormatRegular,
+  TextNumberFormatFilled,
+  TextAlignLeftRegular,
+  TextAlignLeftFilled,
+  CalendarRegular,
+  CalendarFilled,
+  ToggleLeftRegular,
+  ToggleLeftFilled
+} from '@fluentui/react-icons'
+
+/**
+ * An object containing the available Fluent icons and their corresponding regular and filled versions.
+ */
+export const iconCatalog = {
+  ToggleLeft: {
+    regular: ToggleLeftRegular,
+    filled: ToggleLeftFilled
+  },
+  Calendar: {
+    regular: CalendarRegular,
+    filled: CalendarFilled
+  },
+  TextAlignLeft: {
+    regular: TextAlignLeftRegular,
+    filled: TextAlignLeftFilled
+  },
+  TextNumberFormat: {
+    regular: TextNumberFormatRegular,
+    filled: TextNumberFormatFilled
+  },
+  Person: {
+    regular: PersonRegular,
+    filled: PersonFilled
+  },
+  CheckmarkCircle: {
+    regular: CheckmarkCircleRegular,
+    filled: CheckmarkCircleFilled
+  },
+  Checkmark: {
+    regular: CheckmarkRegular,
+    filled: CheckmarkFilled
+  },
+  ConvertRange: {
+    regular: ConvertRangeRegular,
+    filled: ConvertRangeFilled
+  },
+  ArrowSync: {
+    regular: ArrowSyncRegular,
+    filled: ArrowSyncFilled
+  },
+  ArrowSort: {
+    regular: ArrowSort24Regular,
+    filled: ArrowSort24Filled
+  },
+  Grid: {
+    regular: GridRegular,
+    filled: GridFilled
+  },
+  AppsList: {
+    regular: AppsListRegular,
+    filled: AppsListFilled
+  },
+  TextBulletList: {
+    regular: TextBulletListLtrRegular,
+    filled: TextBulletListLtrFilled
+  },
+  TextSortAscending: {
+    regular: TextSortAscendingRegular,
+    filled: TextSortAscendingFilled
+  },
+  TextSortDescending: {
+    regular: TextSortDescendingRegular,
+    filled: TextSortDescendingFilled
+  },
+  TableSettings: {
+    regular: TableSettingsRegular,
+    filled: TableSettingsFilled
+  },
+  GroupList: {
+    regular: GroupListRegular,
+    filled: GroupListFilled
+  },
+  TableCellEdit: {
+    regular: TableCellEditRegular,
+    filled: TableCellEditFilled
+  },
+  ArrowLeft: {
+    regular: ArrowLeftRegular,
+    filled: ArrowLeftFilled
+  },
+  ArrowRight: {
+    regular: ArrowRightRegular,
+    filled: ArrowRightFilled
+  },
+  Eye: {
+    regular: EyeRegular,
+    filled: EyeFilled
+  },
+  Add: {
+    regular: AddRegular,
+    filled: AddFilled
+  }
+}
