+++ SharePointFramework/shared-library/src/icons/index.tsx
@@ -1,128 +1,8 @@
 import { Icon } from '@fluentui/react'
-import {
-  bundleIcon,
-  AppsListFilled,
-  AppsListRegular,
-  GridFilled,
-  GridRegular,
-  TextBulletListLtrFilled,
-  TextBulletListLtrRegular,
-  TextSortAscendingFilled,
-  TextSortAscendingRegular,
-  TextSortDescendingFilled,
-  TextSortDescendingRegular,
-  TableSettingsRegular,
-  TableSettingsFilled,
-  GroupListRegular,
-  GroupListFilled,
-  TableCellEditRegular,
-  TableCellEditFilled,
-  ArrowLeftRegular,
-  ArrowLeftFilled,
-  ArrowRightRegular,
-  ArrowRightFilled,
-  EyeRegular,
-  EyeFilled,
-  AddRegular,
-  AddFilled,
-  ArrowSort24Regular,
-  ArrowSort24Filled,
-  ArrowSyncRegular,
-  ArrowSyncFilled,
-  ConvertRangeRegular,
-  ConvertRangeFilled,
-  CheckmarkRegular,
-  CheckmarkFilled,
-  CheckmarkCircleRegular,
-  CheckmarkCircleFilled,
-} from '@fluentui/react-icons'
+import { bundleIcon } from '@fluentui/react-icons'
 import React, { CSSProperties } from 'react'
-
-/**
- * An object containing the available Fluent icons and their corresponding regular and filled versions.
- */
-const iconCatalog = {
-  CheckmarkCircle: {
-    regular: CheckmarkCircleRegular,
-    filled: CheckmarkCircleFilled
-  },
-  Checkmark: {
-    regular: CheckmarkRegular,
-    filled: CheckmarkFilled
-  },
-  ConvertRange: {
-    regular: ConvertRangeRegular,
-    filled: ConvertRangeFilled
-  },
-  ArrowSync: {
-    regular: ArrowSyncRegular,
-    filled: ArrowSyncFilled
-  },
-  ArrowSort: {
-    regular: ArrowSort24Regular,
-    filled: ArrowSort24Filled
-  },
-  Grid: {
-    regular: GridRegular,
-    filled: GridFilled
-  },
-  AppsList: {
-    regular: AppsListRegular,
-    filled: AppsListFilled
-  },
-  TextBulletList: {
-    regular: TextBulletListLtrRegular,
-    filled: TextBulletListLtrFilled
-  },
-  TextSortAscending: {
-    regular: TextSortAscendingRegular,
-    filled: TextSortAscendingFilled
-  },
-  TextSortDescending: {
-    regular: TextSortDescendingRegular,
-    filled: TextSortDescendingFilled
-  },
-  TableSettings: {
-    regular: TableSettingsRegular,
-    filled: TableSettingsFilled
-  },
-  GroupList: {
-    regular: GroupListRegular,
-    filled: GroupListFilled
-  },
-  TableCellEdit: {
-    regular: TableCellEditRegular,
-    filled: TableCellEditFilled
-  },
-  ArrowLeft: {
-    regular: ArrowLeftRegular,
-    filled: ArrowLeftFilled
-  },
-  ArrowRight: {
-    regular: ArrowRightRegular,
-    filled: ArrowRightFilled
-  },
-  Eye: {
-    regular: EyeRegular,
-    filled: EyeFilled
-  },
-  Add: {
-    regular: AddRegular,
-    filled: AddFilled
-  }
-}
-
-/**
- * Represents the name of a Fluent UI icon.
- */
-export type FluentIconName = keyof typeof iconCatalog
-
-type GetFluentIconOptions = {
-  bundle?: boolean
-  color?: string
-  size?: string | number
-  filled?: boolean
-}
+import { iconCatalog } from './iconCatalog'
+import { FluentIconName, GetFluentIconOptions } from './types'
 
 /**
  * Returns the Fluent icon with the specified name.
@@ -180,3 +60,5 @@ export function getFluentIconWithFallback(name: string, bundleWithFilled = true,
   }
   return <Icon iconName={name} style={{ color }} />
 }
+
+export * from './types'
